Move native build logic into buildSrc
