+++ buildSrc/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    `kotlin-dsl-base`
+}
+
+repositories {
+    google()
+    mavenCentral()
+}
