+++ buildSrc/settings.gradle.kts
@@ -1,0 +1,1 @@
+rootProject.name = "buildSrc"
