+++ buildSrc/src/main/kotlin/com/dokar/quickjs/disableUnsupportedPlatformTasks.kt
@@ -1,0 +1,14 @@
+package com.dokar.quickjs
+
+import org.gradle.api.Project
+
+fun Project.disableUnsupportedPlatformTasks() {
+    tasks.configureEach {
+        val taskName = name.lowercase()
+        if (taskName.contains("linuxx64")) {
+            enabled = currentPlatform == Platform.linux_x64
+        } else if (taskName.contains("mingwx64")) {
+            enabled = currentPlatform == Platform.windows_x64
+        }
+    }
+}
